Add HDF5 client
This adds an HDF5 client that can be used to make HDF5 applications
write their data and metadata to JULEA's object and key-value backends,
respectively.

It uses HDF5's VOL feature and therefore requires a development version
of HDF5 for now. The client can be linked to any application. It can
then be registered and used using the HDF5 API.