JBackend: stateless object backends
This commit makes the object backens stateless. Previously they all
store there data from initialization, mainly a path, in a global
variable. This made it impossible to load and initialization the same
backend twice with different parameters. E.g. two posix backends one on
a SSD and another one on a HDD.

Changes:
jbackend.h:
The JBackend struct were extended by a gpointer called data
to hold arbitrary data specified by each backend.
The signature of the backend functions are extended by a gpointer to
this new data field. The init function is given the address of the new
data field for initialization.

jbackend.c:
The JBackend returned by j_backend_load is a pointer to a
static variabel in the backend and therefore always the same and not
suitable to hold differently initialized instances of a backend. To
overcome this j_object_init know allocates a new JBackend and copys the
data from the JBackend loaded via j_backend_load. On initialization the
backen can then store its data in the new data field of this JBackend.
The new JBackend is free in j_backend_fini.

server.c and jcommon.c:
They both  need to pass the address of JBackend pointer to
j_backend_init now, as j_object_init replaces the data withe the new
initialized one.

Backends:
In all backends the global variable where move to a new JBackendData
struct which is initialized in the init function and free in fini.

posix.c:
Here only the path were move to the JBackendData struct as the other
global variables holding hash tables for caching can not be moved to a
struct easily without altering the previous behavior. This means all
loaded backend are sharing a caches, the global one as well as thread
local cache.

This functionality is need for a future HSM implementation.