lib/hdf5-db: Add workaround for missing ID in new schemas (#113)
When creating a schema, its ID will have the special type J_DB_TYPE_ID,
which is converted to a real type (J_DB_TYPE_UINT32) by the backend. If
we do not get the schema after creating it, the special type will still
be in memory, causing errors when j_db_entry_set_field is called for an
ID field.